+++ pkg/autorenew/token.go
@@ -1,15 +1,15 @@
 package autorenew
 
 import (
+	"context"
 	"fmt"
-	"os"
 
 	vault "github.com/hashicorp/vault/api"
 )
 
 //Token handles the management of the tokens by renewing it
 // and keeping it valid and usable
-func Token(name string, secretWatcher *vault.LifetimeWatcher, interrupt chan os.Signal) {
+func Token(name string, secretWatcher *vault.LifetimeWatcher, ctx context.Context) {
 	fmt.Printf("Strating auto renew of secret %s\n", name)
 
 	// Starting the go routines for managing renewals.
@@ -32,8 +32,8 @@ func Token(name string, secretWatcher *vault.LifetimeWatcher, interrupt chan os.
 		case renew := <-secretWatcher.RenewCh():
 			fmt.Printf("Successfully renewed %s at: %s \n", name, renew.RenewedAt)
 		// Ctrl + c handling.
-		case close := <-interrupt:
-			fmt.Printf("We are closing the renewal of %s %v \n", name, close)
+		case <- ctx.Done():
+			fmt.Printf("We are closing the renewal of %s \n", name)
 			return
 		}
 	}